LayoutTests/webanimations/background-image-css-variable-no-transition-expected.txt

PASS numberOfAnimationsBeforeStyleChange is 0
PASS numberOfAnimationsAfterStyleChange is 0
PASS successfullyParsed is true

TEST COMPLETE
